+++ csvapi/parser.py
@@ -20,7 +20,11 @@ def detect_encoding(filepath):
 
 
 def from_csv(filepath, encoding='utf-8', sniff_limit=SNIFF_LIMIT):
-    return agate.Table.from_csv(filepath, sniff_limit=sniff_limit, encoding=encoding)
+    """Try first w/ sniffing and then w/o sniffing if it fails"""
+    try:
+        return agate.Table.from_csv(filepath, sniff_limit=sniff_limit, encoding=encoding)
+    except ValueError:
+        return agate.Table.from_csv(filepath, encoding=encoding)
 
 
 def from_excel(filepath):